src/pyproject.toml
[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.ruff]
line-length = 120
select = ["D", "B", "S", "A", "C4", "T10", "F", "Q"]
ignore = ["D100", "D107", "D101", "D106", "D103", "D104", "D102", "D105", "D203", "D212", "B008", "B904", "A003"]
exclude = ["__init__.py"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["likeinlife", "maxim-zaytsev", "jannyss", "BenderEg"]

[tool.poetry.dependencies]
python = "^3.10"
django-debug-toolbar = "3.4.0"
django-split-settings = "1.2.0"
django = "5.0"
orjson = "3.9.10"
psycopg2-binary = "2.9.9"
pydantic = "2.5.3"
python-dotenv = "1.0.0"
uwsgi = "2.0.22"
django-structlog = "^7.1.0"
pydantic-settings = "^2.1.0"
djangorestframework = "^3.14.0"
yookassa = "^3.0.0"
dependency-injector = "^4.41.0"
drf-spectacular = "^0.27.1"
celery = "^5.3.6"
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"
mypy = "^1.8.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
